requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
asttokens==2.0.5
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chainlit==1.1.305
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.8
dataclasses-json==0.5.14
debugpy==1.6.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==0.8.3
fastapi==0.110.3
fastapi-cli==0.0.4
fastapi-pagination==0.12.7
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter-core==4.10.0
kubernetes==30.1.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-openai==0.1.9
langchain-text-splitters==0.2.1
langsmith==0.1.81
Lazify==0.4.0
limits==3.12.0
literalai==0.0.607
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.3
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.35.3
openpyxl==3.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.5
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.3
pdfminer.six==20231228
pdfplumber==0.11.1
pickleshare==0.7.5
pillow==10.3.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
posthog==3.5.0
prompt-toolkit==3.0.29
protobuf==4.25.3
psutil==5.9.1
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
pywin32==304
PyYAML==6.0.1
pyzmq==23.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.1
sentence-transformers==2.7.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
slowapi==0.1.9
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.3.0
starlette==0.37.2
streamlit==1.36.0
sympy==1.12.1
syncer==2.0.3
tbb==2021.13.0
tenacity==8.4.1
threadpoolctl==3.4.0
tiktoken==0.7.0
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
tornado==6.1
tqdm==4.66.4
traitlets==5.3.0
transformers==4.34.0
typer==0.12.3
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uptrace==1.24.0
urllib3==2.2.2
uvicorn==0.25.0
watchdog==4.0.1
watchfiles==0.20.0
wcwidth==0.2.5
websocket-client==1.8.0
websockets==12.0
wget==3.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
